add repositorios, dtos y demás cosas

+++ src/main/java/com/eshop/progavanzada/repositories/UserRepository.java
@@ -1,10 +1,9 @@
 package com.eshop.progavanzada.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-// import org.springframework.security.core.userdetails.UserDetails;
 
 import com.eshop.progavanzada.models.User;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
   User findByUsername(String username);
 }
